Refactor out href logic into hook
